backend/requirements.txt
celery==5.5.2
redis==5.0.1
click==8.2.1
Flask==3.1.1
flask_cors==5.0.1
Flask_JWT_Extended==4.7.1
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.41
Werkzeug==3.1.3
psycopg2-binary==2.9.9
